Quick fix: Make undef ID var FAIL in object load
MTObject->load($id) is some seriously dangerous code.  It will either attempt to load one road by ID or, if $id is undef, it will load the latest row in the table.  This amounts to a silent error as it doesn't fail or return nothing, it returns an object of the correct class but THE WRONG OBJECT.  In the worst case scenario, that could cause data corruption.

We SHOULD be doing input checking ALWAYS and returning errors when required variables aren't set.  However, with this commit, I quickly hack together a resolution to the issue (which is EVERYWHERE) in changing all instances of `$class->load($id)` to `$class->load({ id => $id })`. This way it will at least NOT return the wrong object and—if we're lucky—it will blow up, returning an error.